Removed implicit conversion of points for curve functions
Hello,

it might be required to deliver meta information with your value, if you are using an custom curve.
The conversion is not necessary neither, as the d3 provided curve functions already do the trick.

So we are limiting the point function of an curve to number types only.
If you are using an self defined curve class, you will take care of the given data, if not you will use the build in ones that will do.

It would be nice to hear your opinion. :)
Regards!